+++ src/RipTide.cpp
@@ -2255,19 +2255,19 @@ bool GetUpcastType(int numpyInType1, int numpyInType2, int & convertType1, int &
 //  return value 0: one loop can process all data, false = multiple loops
 //  NOTE: if return value is 0 and itemsze == stride, then vector math possible
 //
-int GetStridesAndContig(PyArrayObject const * inArray, int & ndim, int32_t & stride)
+int GetStridesAndContig(PyArrayObject const * inArray, int & ndim, int64_t & stride)
 {
     stride = static_cast<int32_t>(PyArray_ITEMSIZE(inArray));
     int direction = 0;
     ndim = PyArray_NDIM(inArray);
     if (ndim > 0)
     {
-        stride = static_cast<int32_t>(PyArray_STRIDE(inArray, 0));
+        stride = PyArray_STRIDE(inArray, 0);
         if (ndim > 1)
         {
             // at least two strides
             int ndims = PyArray_NDIM(inArray);
-            int32_t lastStride = static_cast<int32_t>(PyArray_STRIDE(inArray, ndims - 1));
+            int64_t lastStride = PyArray_STRIDE(inArray, ndims - 1);
             if (lastStride == stride)
             {
                 // contiguous with one of the dimensions having length 1
@@ -2277,10 +2277,10 @@ int GetStridesAndContig(PyArrayObject const * inArray, int & ndim, int32_t & str
                 // Row Major - 'C' Style
                 // work backwards
                 int currentdim = ndims - 1;
-                int32_t curStrideLen = lastStride;
+                int64_t curStrideLen = lastStride;
                 while (currentdim != 0)
                 {
-                    curStrideLen *= static_cast<int32_t>(PyArray_DIM(inArray, currentdim));
+                    curStrideLen *= PyArray_DIM(inArray, currentdim);
                     LOGGING("'C' %lld vs %lld  dim: %lld  stride: %lld \n", curStrideLen, PyArray_STRIDE(inArray, currentdim - 1),
                             PyArray_DIM(inArray, currentdim - 1), lastStride);
                     if (PyArray_STRIDE(inArray, currentdim - 1) != curStrideLen)
@@ -2294,10 +2294,10 @@ int GetStridesAndContig(PyArrayObject const * inArray, int & ndim, int32_t & str
             {
                 // Col Major - 'F' Style
                 int currentdim = 0;
-                int32_t curStrideLen = stride;
+                int64_t curStrideLen = stride;
                 while (currentdim != (ndims - 1))
                 {
-                    curStrideLen *= static_cast<int32_t>(PyArray_DIM(inArray, currentdim));
+                    curStrideLen *= static_cast<int64_t>(PyArray_DIM(inArray, currentdim));
                     LOGGING("'F' %lld vs %lld  dim:  %lld   stride: %lld \n", curStrideLen,
                             PyArray_STRIDE(inArray, currentdim + 1), PyArray_DIM(inArray, currentdim + 1), stride);
                     if (PyArray_STRIDE(inArray, currentdim + 1) != curStrideLen)
